console: Move pubsub formats logic middleware
