+++ pkg/webui/console/store/middleware/logics/applications.js
@@ -14,7 +14,6 @@
 
 import * as applications from '../../actions/applications'
 import * as link from '../../actions/link'
-import * as pubsubFormats from '../../actions/pubsub-formats'
 
 import api from '../../../api'
 import { isNotFoundError } from '../../../../lib/errors/utils'
@@ -127,14 +126,6 @@ const getApplicationLinkLogic = createRequestLogic({
   },
 })
 
-const getPubsubFormatsLogic = createRequestLogic({
-  type: pubsubFormats.GET_PUBSUB_FORMATS,
-  async process() {
-    const { formats } = await api.application.pubsubs.getFormats()
-    return formats
-  },
-})
-
 export default [
   getApplicationLogic,
   getApplicationDeviceCountLogic,
@@ -142,7 +133,6 @@ export default [
   deleteApplicationLogic,
   getApplicationsLogic,
   getApplicationsRightsLogic,
-  getPubsubFormatsLogic,
   getApplicationLinkLogic,
   ...createEventsConnectLogics(
     applications.SHARED_NAME,
